Add unified timestamps to the logs
Circus captures stdouts and stderrs of managed processes,
so we must specify how to treat with those streams.

Before this commit circus simply wrote the streams to the
apropriate files, so format and presence of timestamps were left
to supervised processes.

This commit will set unified timestamp prefix on applications
streams.